create a job for after install. remove install concern
